added 3scale

+++ models/subscription.py
@@ -4,7 +4,7 @@ from . import PackageManifest
 from . import ClusterServiceVersion
 
 class Subscription:
-    def __init__(self, api:Api, name:str, namespace:str="openshift-operators") -> None:
+    def __init__(self, api:Api, name:str, namespace:str="openshift-operators", skip_operator_group:bool=False) -> None:
         self.api = api
         self.group = "operators.coreos.com"
         self.version = "v1alpha1"
@@ -16,13 +16,17 @@ class Subscription:
         self.namespace = namespace
         self.catalog_source = "redhat-operators"
         self.catalog_source_namespace = "openshift-marketplace"
+        self.cluster_service_version = ""
 
         self.default_operator_group_name = "global-operators"
-        if namespace == "openshift-operators":
-            self.operator_group = OperatorGroup(self.default_operator_group_name, namespace)
+        self.skip_operator_group = skip_operator_group
+        if skip_operator_group:
+            self.operator_group = None
         else:
-            self.operator_group = OperatorGroup(name, namespace)
-        self.cluster_service_version = ""
+            if namespace == "openshift-operators":
+                self.operator_group = OperatorGroup(self.default_operator_group_name, namespace)
+            else:
+                self.operator_group = OperatorGroup(name, namespace)
 
     def get_as_dict(self) -> dict:
         val = {
@@ -49,8 +53,9 @@ class Subscription:
         self.cluster_service_version = package_manifest.get_csv_version(self.channel)
 
         op = self.operator_group
-        if op.name != self.default_operator_group_name:
-            self.api.create_dynamic_object(op.group, op.version, op.kind, op.name, op.namespace, op.get_as_dict())
+        if op is not None:
+            if op.name != self.default_operator_group_name:
+                self.api.create_dynamic_object(op.group, op.version, op.kind, op.name, op.namespace, op.get_as_dict())
         self.api.create_dynamic_object(self.group, self.version, self.kind, self.name, self.namespace, self.get_as_dict())
 
     def destroy(self) -> None:
@@ -58,5 +63,7 @@ class Subscription:
 
         ClusterServiceVersion().destroy_all(self.api, self.name)
         self.api.destroy_dynamic_object(self.group, self.version, self.kind, self.name, self.namespace)
-        if op.name != self.default_operator_group_name:
-            self.api.destroy_dynamic_object(op.group, op.version, op.kind, op.name, op.namespace)
+
+        if op is not None:
+            if op.name != self.default_operator_group_name:
+                self.api.destroy_dynamic_object(op.group, op.version, op.kind, op.name, op.namespace)
